Python/check_db/requirements.txt
asttokens==3.0.0
black==25.1.0
click==8.2.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.1.0
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.14.0
wcwidth==0.2.13
